added weather repo and implemented it

+++ remote/src/main/java/com/jerry/sources/remote/models/CurrentWeatherDTO.kt
@@ -15,20 +15,20 @@ sealed class CurrentWeatherDTO {
         val name: String,
         val region: String,
         val country: String,
-        val lat: Float,
-        val lon: Float,
+        val lat: Double,
+        val lon: Double,
         val tz_id: String,
-        val localtime_epoch: Int,
+        val localtime_epoch: Long,
         val localtime: String
     )
 
     @Serializable
     data class Current(
-        val last_updated_epoch: Int,
+        val last_updated_epoch: Long,
         val last_updated: String,
-        val temp_c: Float,
-        val temp_f: Float,
-        val is_day: Int,
+        val temp_c: Double,
+        val temp_f: Double,
+        val is_day: Long,
         val condition: Condition
     )
 
@@ -36,7 +36,7 @@ sealed class CurrentWeatherDTO {
     data class Condition(
         val text: String,
         val icon: String,
-        val code: Int
+        val code: Long
     )
 
 }
